Kernel disk stats are always in 512-byte units
From kernel documentation:

  The "sectors" in question are the standard UNIX 512-byte
  sectors, not any device- or filesystem-specific block size.

+++ linux/diskmeter.h
@@ -11,7 +11,8 @@
 #include <map>
 #include <string>
 
-typedef std::map<std::string, unsigned long long> diskmap;
+typedef std::map<std::string, unsigned long> diskmap;
+
 
 class DiskMeter : public FieldMeterGraph
     {
